refactor : refactor: 프로젝트 구조 개선을 위한 패키지를 분리한다
- 각 클래스의 역할에 맞춰 프로젝트 구조를 재편성
