+++ src/main/java/racingcar/domain/Application.java
@@ -1,4 +1,11 @@
-package racingcar;
+package racingcar.domain;
+
+import racingcar.controller.GameController;
+import racingcar.domain.strategy.DefaultRandomGenerator;
+import racingcar.domain.strategy.MoveStrategy;
+import racingcar.domain.strategy.RandomMoveStrategy;
+import racingcar.view.InputView;
+import racingcar.view.OutputView;
 
 public class Application {
     public static void main(String[] args) {
