+++ src/main/java/racingcar/domain/RacingGame.java
@@ -1,8 +1,10 @@
-package racingcar;
+package racingcar.domain;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
+import racingcar.domain.strategy.MoveStrategy;
+import racingcar.validator.RacingGameValidator;
 
 public class RacingGame {
     private final List<Car> cars;
