+++ src/main/java/racingcar/domain/strategy/DefaultRandomGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain.strategy;
 
 import camp.nextstep.edu.missionutils.Randoms;
