+++ src/main/java/racingcar/domain/strategy/MoveStrategy.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain.strategy;
 
 public interface MoveStrategy {
     boolean isMovable();
