+++ src/main/java/racingcar/domain/strategy/RandomGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain.strategy;
 
 public interface RandomGenerator {
     int generate();
