+++ src/main/java/racingcar/validator/BaseValidator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.validator;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
