+++ src/main/java/racingcar/validator/RacingGameValidator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.validator;
 
 public class RacingGameValidator extends BaseValidator {
