+++ src/main/java/racingcar/view/InputView.java
@@ -1,10 +1,11 @@
-package racingcar;
+package racingcar.view;
 
 import camp.nextstep.edu.missionutils.Console;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
+import racingcar.validator.InputValidator;
 
 public class InputView {
